requirements.txt
pandas>=1.5
matplotlib>=3.6
pyyaml>=6.0
tqdm>=4.65
geoip2>=4.6   
